+++ .ycm_extra_conf.py
@@ -53,6 +53,8 @@ flags = [
 '-DNDEBUG',
 '-D_POSIX_C_SOURCE=200809L',
 '-D_BSD_SOURCE',
+'-D_GNU_SOURCE',
+'-DSK_DUMP_CORE',
 # THIS IS IMPORTANT! Without a "-std=<something>" flag, clang won't know which
 # language to use when compiling headers. So it will guess. Badly. So C++
 # headers will be compiled as C headers. You don't want that so ALWAYS specify