+++ src/api/sk_eventloop.h
@@ -1,7 +1,7 @@
 #ifndef SK_EVENTLOOP_H
 #define SK_EVENTLOOP_H
 
-void* sk_eventloop_create();
+void* sk_eventloop_create(int max_fds);
 void sk_eventloop_destroy(void* evlp);
 
 // return - > 0: how many event have been processed