+++ src/entities/sk_entity_util.c
@@ -80,9 +80,9 @@ void sk_entity_util_unpack(fev_state* fev, fev_buff* evbuff,
     sk_metrics_global.request.inc(1);
 }
 
-void sk_entity_safe_destroy(sk_entity_t* entity)
+int sk_entity_safe_destroy(sk_entity_t* entity)
 {
-    if (!entity) return;
+    if (!entity) return 0;
     sk_print("sk_entity_safe_destroy: entity %p\n", (void*)entity);
 
     sk_entity_mgr_t* owner = sk_entity_owner(entity);
@@ -92,12 +92,19 @@ void sk_entity_safe_destroy(sk_entity_t* entity)
 
         sk_entity_mark(entity, SK_ENTITY_DEAD);
         sk_entity_destroy(entity);
+        return 0;
     } else {
         sk_sched_t* sched  = SK_ENV_SCHED;
         sk_sched_t* target = sk_entity_mgr_sched(owner);
 
-        sk_sched_send(sched, target, entity, NULL,
+        if (sched != target) {
+            sk_sched_send(sched, target, entity, NULL,
                       SK_PTO_ENTITY_DESTROY, NULL, 0);
+            return 1;
+        } else {
+            sk_entity_mgr_del(owner, entity);
+            return 0;
+        }
     }
 }