+++ src/user-c/src/ep.c
@@ -116,8 +116,14 @@ ep_job_t* _ep_job_create(const skull_service_t* service,
 
     // TODO: Tricky here, should manually convert the fields one by one
     memcpy(sk_handler, handler, sizeof(*handler));
-    sk_handler->unpack  = _unpack;
+    sk_handler->unpack  = job->handler.unpack ? _unpack : NULL;
     sk_handler->release = _release;
+
+    // Hacky here, we force to disable the 'SK_EP_F_CONCURRENT' due to it's a
+    //  too dangerous flag which will can cause the whole ep pool be blocked,
+    //  So let's just force use the ping-pong mode
+    // Also, we disable all other flags, to add more flags in the futher
+    sk_handler->flags = 0;
     return job;
 }