+++ src/user-cpp/include/skullcpp/ep.h
@@ -62,10 +62,6 @@ public:
     (skullcpp::EPClient::EpNPCb) \
     std::bind(f, std::placeholders::_1, std::placeholders::_2, ##__VA_ARGS__)
 
-// EP Handler Flags
-#define SKULLCPP_EP_F_CONCURRENT 0x1
-#define SKULLCPP_EP_F_ORPHAN     0x2
-
 public:
     EPClient();
     ~EPClient();
@@ -76,12 +72,14 @@ public:
     void setIP(const std::string& ip);
 
     // unit: millisecond
-    // <= 0: means no timeout (Only available when 'SKULLCPP_EP_F_ORPHAN' be set)
+    // <= 0: means no timeout
     // >  0: after x milliseconds, the ep call would time out
     void setTimeout(int timeout);
 
-    // Set value of 'SKULLCPP_EP_F_CONCURRENT' and 'SKULLCPP_EP_F_ORPHAN'
+    // Reserved, no effect currently
     void setFlags(int flags);
+
+    // Setup a unpack function
     void setUnpack(UnpackFn unpackFn);
 
 public: