+++ tests/cases/ep_timeout0_responsed/service.cpp
@@ -68,15 +68,8 @@ void skull_service_getdata(const skullcpp::Service& service,
     epClient.setTimeout(0);
     epClient.setUnpack(skull_BindEpUnpack(_ep_unpack));
 
-    // This would be failed
     skullcpp::EPClient::Status st =
         epClient.send(service, "hello ep", skull_BindEpCb(_ep_cb));
-    std::cout << "ep status: " << st << std::endl;
-    SKULLCPP_LOG_INFO("svc.test-get-2", "ep status: " << st);
-
-    // Set flag = SKULLCPP_EP_F_ORPHAN, and try again
-    epClient.setFlags(SKULLCPP_EP_F_ORPHAN);
-    st = epClient.send(service, "hello ep", skull_BindEpCb(_ep_cb));
     std::cout << "ep status1: " << st << std::endl;
     SKULLCPP_LOG_INFO("svc.test-get-3", "ep status: " << st);
 }